+++ saphyr/src/annotated.rs
@@ -19,8 +19,10 @@
 
 pub mod marked_yaml;
 pub mod yaml_data;
+pub mod yaml_data_owned;
 
 pub use yaml_data::{AnnotatedMapping, AnnotatedSequence, AnnotatedYamlIter, YamlData};
+pub use yaml_data_owned::YamlDataOwned;
 
 /// A trait allowing for introspection in the hash types of the [`YamlData::Mapping`] variant.
 ///
@@ -35,6 +37,24 @@ pub trait AnnotatedNode: std::hash::Hash + std::cmp::Eq {
     type HashKey<'a>: From<YamlData<'a, Self::HashKey<'a>>>
         + for<'b> std::cmp::PartialEq<Self::HashKey<'b>>
         + AnnotatedNode;
+
+    /// See [`YamlData::parse_representation_recursive`].
+    fn parse_representation_recursive(&mut self) -> bool;
+}
+
+/// A trait allowing for introspection in the hash types of the [`YamlData::Mapping`] variant.
+///
+/// This trait must be implemented by annotated YAML objects.
+///
+/// See [`LoadableYamlNode::HashKey`] for more details.
+///
+/// [`LoadableYamlNode::HashKey`]: crate::loader::LoadableYamlNode::HashKey
+#[allow(clippy::module_name_repetitions)]
+pub trait AnnotatedNodeOwned: std::hash::Hash + std::cmp::Eq {
+    /// The type used as the key in the [`YamlDataOwned::Mapping`] variant.
+    type HashKey: From<YamlDataOwned<Self::HashKey>>
+        + std::cmp::PartialEq<Self::HashKey>
+        + AnnotatedNodeOwned;
 
     /// See [`YamlData::parse_representation_recursive`].
     fn parse_representation_recursive(&mut self) -> bool;